+++ src/app/style/global.scss
@@ -9,8 +9,3 @@ a {
   text-decoration: none;
   cursor: pointer;
 }
-
-.container {
-  max-width: 1230px;
-  margin: 0 auto;
-}
