+++ src/shared/ui/wrap/index.tsx
@@ -1,0 +1,13 @@
+import React, {FC, PropsWithChildren} from 'react';
+
+import cl from "./style.module.scss"
+
+const Wrap: FC<PropsWithChildren> = ({children}) => {
+  return (
+    <div className={cl.wrap}>
+      <div>{children}</div>
+    </div>
+  );
+};
+
+export default Wrap;
