+++ src/widgets/header/lib/constants.ts
@@ -1,0 +1,27 @@
+type Link = {
+  name: string
+  href: string
+}
+
+export const links: Link[] = [
+  {
+    name: 'About Us',
+    href: '/about',
+  },
+  {
+    name: 'Services',
+    href: '/services',
+  },
+  {
+    name: 'Work',
+    href: '/work',
+  },
+  {
+    name: 'News',
+    href: '/news',
+  },
+  {
+    name: 'Contacts',
+    href: '/contacts',
+  },
+]
